Move registry into folder, split types into file
